fix eliminar categoria

+++ app.py
@@ -257,21 +257,23 @@ def eliminar_mensaje(id_mensaje):
 @app.route('/eliminar_categoria/<id_categoria>')
 @login_required
 def eliminar_categoria(id_categoria):
-    sql="delete from articulo where id_categoria = %s;"
-    conn = mysql.connect()
-    cursor = conn.cursor()
-    datos = (id_categoria)
-    cursor.execute(sql,(datos))
-    conn.commit()
+    if session['loggedin'] == True & session['id_administrador'] == 1 :
+        sql="delete from articulo where id_categoria = %s;"
+        conn = mysql.connect()
+        cursor = conn.cursor()
+        datos = (id_categoria)
+        cursor.execute(sql,(datos))
+        conn.commit()
 
-    sql="delete from categoria where id_categoria = %s;"
+        sql="delete from categoria where id_categoria = %s;"
 
-    datos = (id_categoria)
-    cursor.execute(sql,(datos))
-    conn.commit()
+        datos = (id_categoria)
+        cursor.execute(sql,(datos))
+        conn.commit()
 
-    return redirect(url_for('panel_admin'))
-
+        return redirect(url_for('panel_admin'))
+    else:
+        return redirect(url_for('index'))
 
 
 # funcion de eliminar articulo
